requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.6
django-cors-headers==2.1.0
django-crispy-forms==1.7.0
django-filter==1.1.0
django-leaflet==0.22.0
django-markdown-deux==1.0.5
django-pagedown==0.1.3
django-rest-framework==0.1.0
djangorestframework==3.6.4
djangorestframework-gis==0.11.2
djangorestframework-jwt==1.11.0
flake8==3.5.0
idna==2.6
Markdown==2.6.9
markdown2==2.3.4
mccabe==0.6.1
olefile==0.44
pew==1.1.0
Pillow==4.3.0
pipenv==8.2.7
psutil==5.3.1
psycopg2==2.6.2
pycodestyle==2.3.1
pyflakes==1.6.0
PyJWT==1.5.3
pytz==2017.2
requests==2.18.4
six==1.11.0
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
